Assets/requirements.txt
altgraph==0.17.4
audioread==3.0.1
azure-cognitiveservices-speech==1.41.1
azure-core==1.32.0
azure-identity==1.19.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cryptography==44.0.0
cycler==0.12.1
decorator==5.1.1
filelock==3.13.4
firebase-admin==6.6.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
httplib2==0.22.0
huggingface-hub==0.26.5
HyperPyYAML==1.2.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
joblib==1.4.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mkl==2021.4.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.0.8
namex==0.0.8
networkx==3.3
numba==0.59.1
numpy==1.26.4
package_name==0.1
packaging==24.0
pandas==2.2.2
pefile==2023.2.7
pillow==10.3.0
platformdirs==4.2.1
pooch==1.8.1
portalocker==2.10.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
PyJWT==2.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.1
regex==2024.11.6
requests==2.31.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
setuptools==69.5.1
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
speechbrain==1.0.0
sympy==1.13.1
tbb==2021.12.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
threadpoolctl==3.4.0
tokenizers==0.21.0
torch==2.5.1+cpu
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
tqdm==4.66.2
transformers==4.47.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vboxapi==1.0
Wave==0.0.2
Werkzeug==3.0.3
wget==3.2
wheel==0.43.0
wrapt==1.16.0
